Improve the robustness of the wake sequence
This does two things:

1. Always sleeps the device after a successful or failed request
2. Retries failed wakeups

Previously if a request errorred out and another request was sent close
enough in time that the device hadn't automatically gone to sleep, that
second request would fail. By always sleeping after a request, the
wakeup sequence should always be in a known state and work.
Additionally, just in case, the wakeup sequence does one retry. This
should never be exercised, but feels like the right thing to do.