Serialize I2C requests to the same device
This prevents multiple processes that are using the ATECC508A from
stepping on each other. The symptoms are usually that one process
crashes since the wakeup message will timeout if the chip is busy.

This commit refactors the I2C transport to start a named GenServer for
each unique ATECC508A. Requests are serialized through that GenServer.
The GenServer is supervised in the ATECC508A's supervision tree.